set_cfg_frames_to_be_encoded: cast to int64_t

+++ Source/App/EncApp/EbAppConfig.c
@@ -368,7 +368,7 @@ static void set_cfg_source_height(const char *value, EbConfig *cfg) {
     cfg->config.source_height = strtoul(value, NULL, 0);
 };
 static void set_cfg_frames_to_be_encoded(const char *value, EbConfig *cfg) {
-    cfg->frames_to_be_encoded = strtol(value, NULL, 0);
+    cfg->frames_to_be_encoded = (int64_t)strtol(value, NULL, 0);
 };
 static void set_buffered_input(const char *value, EbConfig *cfg) {
     cfg->buffered_input = strtol(value, NULL, 0);
